refactor: move PrimitiveEffects into phase.typer.jvm
Moves PrimitiveEffects and its four JSON effect tables next to the other Java
interop code in phase.typer.jvm and updates the resource paths accordingly.
TypeReduction2 is its only caller.
