+++ main/src/ca/uwaterloo/flix/language/phase/typer/jvm/PrimitiveEffects.Constructors.json
